Decompose rust_admin auth.rs into cohesive submodules
auth/ now holds handlers (login/refresh/me/logout), tokens (JWT claims
and refresh/CSRF token generation), sessions (refresh session
persistence and rotation), extract (require_admin/require_csrf and
header/cookie token readers), and cookies (Set-Cookie construction).
Pure move with module-root glob re-exports; cookie name/path constants
stay at the module root.
